Split API to other project. Code cleanup. Start of making global services
