+++ Kits/Cooldowns/Models/KitCooldownData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Kits.Cooldowns.Models;
+
+public class KitCooldownData
+{
+    public string? KitName { get; set; }
+    public DateTime KitCooldown { get; set; }
+}
