+++ Kits/Databases/DataStore/KitsData.cs
@@ -1,0 +1,9 @@
+﻿using Kits.API.Models;
+using System.Collections.Generic;
+
+namespace Kits.Databases.DataStore;
+
+public class KitsData
+{
+    public List<Kit>? Kits { get; set; }
+}
