+++ Kits/Databases/DataStoreKitDataStore.cs
@@ -1,0 +1,118 @@
+﻿using Autofac;
+using Kits.API.Databases;
+using Kits.API.Models;
+using Kits.Databases.DataStore;
+using OpenMod.API;
+using OpenMod.API.Commands;
+using OpenMod.Core.Helpers;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Kits.Databases;
+
+public class DataStoreKitDataStore : KitDataStoreCore, IKitDatabaseProvider, IDisposable
+{
+    private const string c_KitsKey = "kits";
+
+    private KitsData m_Data = null!;
+    private IDisposable? m_FileWatcher;
+
+    public DataStoreKitDataStore(ILifetimeScope lifetimeScope) : base(lifetimeScope)
+    {
+    }
+
+    public async Task<bool> AddKitAsync(Kit kit)
+    {
+        if (kit is null)
+        {
+            throw new ArgumentNullException(nameof(kit));
+        }
+
+        if (m_Data.Kits.Any(x => x.Name.Equals(kit.Name, StringComparison.OrdinalIgnoreCase)))
+        {
+            throw new UserFriendlyException(StringLocalizer["commands:kit:exist"]);
+        }
+
+        m_Data.Kits?.Add(kit);
+        await SaveToDisk();
+        return true;
+    }
+
+    public Task<Kit?> FindKitByNameAsync(string name)
+    {
+        return Task.FromResult(m_Data.Kits?.Find(x => x.Name.Equals(name, StringComparison.OrdinalIgnoreCase)));
+    }
+
+    public Task<IReadOnlyCollection<Kit>> GetKitsAsync()
+    {
+        return Task.FromResult((IReadOnlyCollection<Kit>)(m_Data.Kits ?? new()));
+    }
+
+    public async Task LoadDatabaseAsync()
+    {
+        await LoadFromDisk();
+
+        var component = LifetimeScope.Resolve<IOpenModComponent>();
+
+        m_FileWatcher = DataStore.AddChangeWatcher(c_KitsKey, component,
+            () => AsyncHelper.RunSync(LoadFromDisk));
+
+        await SaveToDisk();
+    }
+
+    private async Task LoadFromDisk()
+    {
+        if (await DataStore.ExistsAsync(c_KitsKey))
+        {
+            m_Data = await DataStore.LoadAsync<KitsData>(c_KitsKey) ?? new();
+            m_Data.Kits ??= new();
+            return;
+        }
+
+        m_Data = new() { Kits = new() };
+    }
+
+    public async Task<bool> RemoveKitAsync(string name)
+    {
+        var index = m_Data.Kits?.FindIndex(x => x.Name.Equals(name, StringComparison.OrdinalIgnoreCase));
+        if (index < 0)
+        {
+            throw new UserFriendlyException(StringLocalizer["commands:kit:remove:fail", new { Name = name }]);
+        }
+
+        m_Data.Kits?.RemoveAt(index!.Value);
+        await SaveToDisk();
+        return true;
+    }
+
+    public async Task<bool> UpdateKitAsync(Kit kit)
+    {
+        if (kit is null)
+        {
+            throw new ArgumentNullException(nameof(kit));
+        }
+
+        var index = m_Data.Kits?.FindIndex(
+            x => x.Name.Equals(kit.Name, StringComparison.OrdinalIgnoreCase));
+        if (index < 0)
+        {
+            return false;
+        }
+
+        m_Data.Kits![index!.Value] = kit;
+        await SaveToDisk();
+        return true;
+    }
+
+    public void Dispose()
+    {
+        m_FileWatcher?.Dispose();
+    }
+
+    private Task SaveToDisk()
+    {
+        return DataStore.SaveAsync(c_KitsKey, m_Data);
+    }
+}
