+++ Kits/Extensions/ConvertorExtension.cs
@@ -1,4 +1,4 @@
-﻿using Kits.API;
+﻿using Kits.API.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
@@ -7,86 +7,76 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace Kits.Extensions
+namespace Kits.Extensions;
+
+internal static class ConvertorExtension
 {
-    public static class ConvertorExtension
+    public static readonly byte s_SaveVersion = 1;
+
+    public static KitItem ConvertIItemToKitItem(this IItem item)
     {
-        public static readonly byte s_SaveVersion = 1;
+        return new(item.Asset.ItemAssetId, item.State);
+    }
 
-        public static KitItem ConvertIItemToKitItem(this IItem item)
+    public static byte[] ConvertToByteArray(this List<KitItem>? items)
+    {
+        if (items == null)
         {
-            return new(item.Asset.ItemAssetId, item.State);
+            return Array.Empty<byte>();
         }
 
-        public static byte[] ConvertToByteArray(this List<KitItem>? items)
+        using var ms = new MemoryStream();
+        using var bw = new BinaryWriter(ms);
+
+        bw.Write(s_SaveVersion);
+        bw.Write(items.Count);
+
+        foreach (var item in items)
         {
-            if (items == null)
-            {
-                return Array.Empty<byte>();
-            }
-
-            using var ms = new MemoryStream();
-            using var bw = new BinaryWriter(ms);
-
-            bw.Write(s_SaveVersion);
-            bw.Write(items.Count);
-
-            foreach (var item in items)
-            {
-                item.Serialize(bw);
-            }
-
-            return ms.ToArray();
+            item.Serialize(bw);
         }
 
-        public static List<KitItem> ConvertToKitItems(this byte[]? block)
+        return ms.ToArray();
+    }
+
+    public static List<KitItem>? ConvertToKitItems(this byte[]? block)
+    {
+        if (block == null)
         {
-            var output = new List<KitItem>();
-
-            if (block == null)
-            {
-                return output;
-            }
-
-            using var ms = new MemoryStream(block, false);
-            using var br = new BinaryReader(ms);
-
-            br.ReadByte(); // save version, for now ignored
-
-            var count = br.ReadInt32();
-            for (var i = 0; i < count; i++)
-            {
-                var kitItem = new KitItem();
-                kitItem.Deserialize(br);
-
-                output.Add(kitItem);
-            }
-
-            return output;
+            return null;
         }
 
-        // https://stackoverflow.com/questions/44829824/how-to-store-json-in-an-entity-field-with-ef-core
-        public static PropertyBuilder<List<KitItem>?> HasByteArrayConversion(this PropertyBuilder<List<KitItem>?> propertyBuilder)
+        var output = new List<KitItem>();
+
+        using var ms = new MemoryStream(block, false);
+        using var br = new BinaryReader(ms);
+
+        br.ReadByte(); // save version, for now ignored
+
+        var count = br.ReadInt32();
+        for (var i = 0; i < count; i++)
         {
-            var converter = new ValueConverter<List<KitItem>?, byte[]>
-            (
-            v => v.ConvertToByteArray(),
-            v => v.ConvertToKitItems()
-            );
+            var kitItem = new KitItem();
+            kitItem.Deserialize(br);
 
-            /* var comparer = new ValueComparer<List<KitItem>?>
-             (
-                 (l, r) => l == r,
-                 v => v == null ? 0 : v.GetHashCode(),
-                 v => v.ConvertToByteArray().ConvertToKitItems()
-             );*/
-
-            propertyBuilder.HasConversion(converter);
-            propertyBuilder.Metadata.SetValueConverter(converter);
-            //propertyBuilder.Metadata.SetValueComparer(comparer);
-            //propertyBuilder.HasColumnType("longblob");
-
-            return propertyBuilder;
+            output.Add(kitItem);
         }
+
+        return output;
+    }
+
+    // https://stackoverflow.com/questions/44829824/how-to-store-json-in-an-entity-field-with-ef-core
+    public static PropertyBuilder<List<KitItem>?> HasByteArrayConversion(this PropertyBuilder<List<KitItem>?> propertyBuilder)
+    {
+        var converter = new ValueConverter<List<KitItem>?, byte[]>
+        (
+        v => v.ConvertToByteArray(),
+        v => v.ConvertToKitItems()
+        );
+
+        propertyBuilder.HasConversion(converter);
+        propertyBuilder.Metadata.SetValueConverter(converter);
+
+        return propertyBuilder;
     }
 }
