--- Kits/Models/KitCooldownData.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Kits.Models
-{
-    public class KitCooldownData
-    {
-        public string? KitName { get; set; }
-        public DateTime KitCooldown { get; set; }
-    }
-}
