--- Kits/Models/KitsCooldownData.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Kits.Models
-{
-    public class KitsCooldownData
-    {
-        public Dictionary<string, List<KitCooldownData>>? KitsCooldown { get; set; }
-    }
-}
