--- Kits/Models/KitsData.cs
@@ -1,10 +1,0 @@
-﻿using Kits.API;
-using System.Collections.Generic;
-
-namespace Kits.Models
-{
-    public class KitsData
-    {
-        public List<Kit>? Kits { get; set; }
-    }
-}
